UNDO: Support for the <guid> RSS element
Some feeds contain episodes whose URL changes with time. (For example, to
archive old episodes to a different server) Unfortunately, hpodder will see
this as a new episode and re-download it.

This patch treats the <guid> RSS element, if present (the <guid> element is
optional), as an additional unique identifier for an episode. If the URL for an
episode with a GUID changes, the URL of the old episode will be updated and the
episode will not be re-downloaded.